pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "distributed-rate-limiter"
version = "0.1.0"
description = "Distributed rate limiting service with FastAPI, Redis, PostgreSQL, and Prometheus."
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13.2",
  "asyncpg>=0.30.0",
  "fastapi>=0.115.0",
  "httpx>=0.28.1",
  "prometheus-client>=0.21.0",
  "pydantic-settings>=2.6.1",
  "redis[hiredis]>=5.2.0",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
  "asgi-lifespan>=2.1.0",
  "locust>=2.32.4",
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.7.1",
  "testcontainers>=4.8.2",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
  "unit: fast pure-python unit tests",
  "integration: tests that exercise Redis/PostgreSQL-backed flows",
]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "N", "ASYNC"]

